Add round-up pagination navigation
This allows navigating to the next multiple of the page size when close
to the end of the list. For example, given the last 10 entries out of 11
at an offset of 1, the database query will not find any *more* entries,
but one should still be able to navigate to page 2 (consisting of the
very last entry).