Prohibit Friendship interactions with oneself
While the idea of being Friends with oneself is not *entirely*
meaningless (it could, say, relate to one's self-esteem), banning
it should avoid any confusion on that matter.

Should self-Friendship be allowed, the check for
"reverse_row['status'] < 0 and new_status > 0" may need to be
adjusted to consider this special case -- or deliberately left as-is
(for entertainment).